Updated: - successfully finished folder structure creation/ download s3 file step
