+++ funcs.py
@@ -4,194 +4,414 @@ import pandas as pd
 import logging
 import re
 from subprocess import Popen, PIPE, STDOUT
+import constants_and_names as cn
 
 #TODO: Make sure masks are being written out to their correct folders. Include one with mangrove, and one without?
 
-TCD_THRESHOLD = 30
-GAIN_FLAG = True
-MASK_PATHS = {"IDN": "00N_110E", "GMB": "20N_20W"}
-
-def get_tile_id(raster_name):
-    return raster_name.split("_", 3)[3]
-
-def get_mask_tiles(tile_id):
-    if "00N_110E" in tile_id:
-        return os.path.join(arcpy.env.workspace, "Mask", "Mask", "00N_110E")
+def DownloadFiles(working_directory):
+    # Step 1: Checking to see if the AOIS folder exists and if it contains a shapefile
+    print("Step 0.1: Checking to see if AOIS folder exists and contains a shapefile")
+    if os.path.isdir(cn.aois_folder):
+        print(f"    Success: {cn.aois_folder} exists.")
+        # Checking to see if the AOIS folder has any shapefiles
+        aoi_list = [os.path.join(cn.aois_folder, f) for f in os.listdir(cn.aois_folder) if f.endswith('.shp')]
+        if len(aoi_list) >= 1:
+            print(f"    Success: {cn.aois_folder} contains {len(aoi_list)} shapefiles.")
+        else:
+            raise Exception(f"  Failure: {cn.aois_folder} does not contain a shapefile.")
     else:
-        return os.path.join(arcpy.env.workspace, "Mask", "Mask", "20N_20W")
+        raise Exception(f"  Failure: {cn.aois_folder} does not exist.")
 
-def process_raster(raster, gain_tiles, mangrove_tiles, whrc_tiles, mask_tiles):
-    tile_id = get_tile_id(os.path.splitext(os.path.basename(raster))[0])
+    # Step 2: Create Input folder and subfolder for each tile in tile_list
+    print("Step 0.2: Creating Input folder structure")
+    for tile in cn.tile_list:
+        tile_id_folder = os.path.join(cn.input_folder, tile)
+        folder_check(tile_id_folder)
 
-    gain_raster_path = os.path.join(gain_tiles, f'{tile_id}_gain.tif')
-    mangrove_raster_path = os.path.join(mangrove_tiles, f'{tile_id}_mangrove_agb_t_ha_2000_rewindow.tif')
-    whrc_raster_path = os.path.join(whrc_tiles, f'{tile_id}_t_aboveground_biomass_ha_2000.tif')
+    # Step 3: Create Mask folder with Inputs folder (contains gain, mangrove, pre-200 plantations, and tree cover density subfolders) and Mask folder (contains subfolders for each tile in tile_list)
+    print("Step 0.3: Creating Mask folder structure")
+    for subfolder in [cn.mask_input_folder, cn.gain_folder, cn.mangrove_folder, cn.plantations_folder, cn.tcd_folder,
+                      cn.whrc_folder]:
+        folder_check(subfolder)
+    for tile in cn.tile_list:
+        tile_id_folder = os.path.join(cn.mask_output_folder, tile)
+        folder_check(tile_id_folder)
+    # TODO: Update code so that the mask output gets written to the tile folder rather than the Mask, Mask folder
 
-    mask_path_tcd_gain = os.path.join(mask_tiles, f'{tile_id}_tcd_gain.tif')
-    print(f'Creating masks for {tile_id}: \n')
+    # Step 4: Create Output folder with Annual folder, CSV folder, and subfolder for each tile in tile_list
+    print("Step 0.4: Creating Output folder structure")
+    for subfolder in [cn.csv_folder, cn.annual_folder]:
+        folder_check(subfolder)
+    for tile in cn.tile_list:
+        tile_id_folder = os.path.join(cn.outputs_folder, tile)
+        folder_check(tile_id_folder)
 
-    tcd_raster = arcpy.sa.Con(arcpy.Raster(raster) > TCD_THRESHOLD, 1, 0)
-    gain_raster = arcpy.sa.Con(arcpy.Raster(gain_raster_path) > 0, 1, 0)
-    whrc_raster = arcpy.sa.Con(arcpy.Raster(whrc_raster_path) > 0, 1, 0)
-    mangrove_raster = arcpy.sa.Con(arcpy.Raster(mangrove_raster_path) > 0, 1, 0)
+    # Step 5: Create TCL folder structure
+    print("Step 0.5: Creating TCL folder structure")
+    folder_check(cn.tcl_folder)
+    # TODO: Where are these clipped rasters coming from?
 
-    output_raster = arcpy.sa.Times(tcd_raster, whrc_raster)
-    output_raster = arcpy.sa.Plus(output_raster, gain_raster)
-    output_raster = arcpy.sa.Plus(output_raster, mangrove_raster)
-    output_raster = arcpy.sa.Con(output_raster > 0, 1, 0)
+    # Step 6: Download emission/removal/netflux tiles (6 per tile) to Input folder
+    print("Step 0.6: Downloading files for Input folder")
+    s3_flexible_download(cn.tile_list, cn.gross_emis_forest_extent_s3_path, cn.gross_emis_forest_extent_s3_pattern,
+                         cn.input_folder)
+    s3_flexible_download(cn.tile_list, cn.gross_emis_full_extent_s3_path, cn.gross_emis_full_extent_s3_pattern,
+                         cn.input_folder)
+    s3_flexible_download(cn.tile_list, cn.gross_removals_forest_extent_s3_path,
+                         cn.gross_removals_forest_extent_s3_pattern, cn.input_folder)
+    s3_flexible_download(cn.tile_list, cn.gross_removals_full_extent_s3_path, cn.gross_removals_full_extent_s3_pattern,
+                         cn.input_folder)
+    s3_flexible_download(cn.tile_list, cn.netflux_forest_extent_s3_path, cn.netflux_forest_extent_s3_pattern,
+                         cn.input_folder)
+    s3_flexible_download(cn.tile_list, cn.netflux_full_extent_s3_path, cn.netflux_full_extent_s3_pattern,
+                         cn.input_folder)
 
-    output_raster.save(mask_path_tcd_gain)
+    # Step 7: Download Gain, Mangrove, Pre_2000_Plantations, TCD, and WHRC subfolders for each tile to Mask, Inputs subfolders
+    print("Step 0.6: Downloading files for Mask/Inputs folder")
+    s3_flexible_download(cn.tile_list, cn.gain_s3_path, cn.gain_s3_pattern, cn.gain_folder, cn.gain_local_pattern)
+    s3_flexible_download(cn.tile_list, cn.mangrove_s3_path, cn.mangrove_s3_pattern, cn.mangrove_folder)
+    s3_flexible_download(cn.tile_list, cn.plantation_s3_path, cn.plantation_s3_pattern, cn.plantations_folder)
+    s3_flexible_download(cn.tile_list, cn.tcd_s3_path, cn.tcd_s3_pattern, cn.tcd_folder)
+    s3_flexible_download(cn.tile_list, cn.whrc_s3_path, cn.whrc_s3_pattern, cn.whrc_folder)
 
-def create_masks():
+    # Step 8: Download TCL folder tiles
+
+def CreateMasks(input_folder):
     mask_inputs = os.path.join(arcpy.env.workspace, "Mask", "Inputs")
+    mask_tiles = os.path.join(arcpy.env.workspace, "Mask","Mask")
 
     tcd_tiles = os.path.join(mask_inputs, "TCD")
     gain_tiles = os.path.join(mask_inputs, "Gain")
     whrc_tiles = os.path.join(mask_inputs, "WHRC")
     mangrove_tiles = os.path.join(mask_inputs, "Mangrove")
 
+    TDC_threshold = 30
+    gain = True
+
+    # Get a list of rasters in the raster folder
     tcd_list = [os.path.join(tcd_tiles, f) for f in os.listdir(tcd_tiles) if f.endswith('.tif')]
     print(tcd_list)
 
+    # Loop through each raster and calculate zonal statistics as table
     for raster in tcd_list:
-        tile_id = get_tile_id(os.path.splitext(os.path.basename(raster))[0])
-        mask_tiles = get_mask_tiles(tile_id)
-        process_raster(raster, gain_tiles, mangrove_tiles, whrc_tiles, mask_tiles)
-
-def list_files_in_directory(directory, file_extension):
-    return [os.path.join(directory, f) for f in os.listdir(directory) if f.endswith(file_extension)]
-
-def process_zonal_statistics(aoi, raster_folder, output_folder):
-    raster_list = list_files_in_directory(raster_folder, '.tif')
-
-    for raster in raster_list:
+        # Extract the name of the raster from the file path
         raster_name = os.path.splitext(os.path.basename(raster))[0]
-        print(f'Calculating zonal statistics for {raster_name}: \n')
+        print(raster_name)
+        tile_id = raster_name.split("_", 3)[3]
+        print(tile_id)
+        gain_raster = os.path.join(gain_tiles, str(tile_id + '_gain.tif'))
+        mangrove_raster = os.path.join(mangrove_tiles, str(tile_id + '_mangrove_agb_t_ha_2000_rewindow.tif'))
+        whrc_raster = os.path.join(whrc_tiles, str(tile_id + '_t_aboveground_biomass_ha_2000.tif'))
 
-        output_name = f"{os.path.splitext(os.path.basename(aoi))[0]}_{raster_name}.dbf"
-        output_path = os.path.join(output_folder, output_name)
+        mask_path_tcd = os.path.join(mask_tiles, str(tile_id + '_tcd.tif'))
+        mask_path_tcd_gain = os.path.join(mask_tiles, str(tile_id + '_tcd_gain.tif'))
+        print('Creating masks for ' + tile_id + ": \n")
 
-        arcpy.gp.ZonalStatisticsAsTable_sa(aoi, "GID_0", raster, output_path, "DATA", "SUM")
-        csv_file = f"{os.path.splitext(os.path.basename(aoi))[0]}_{raster_name}.csv"
-        arcpy.TableToTable_conversion(output_path, output_folder, csv_file)
+        # Conditional logic for TCD
+        tcd_raster = arcpy.sa.Con(arcpy.Raster(raster) > 30, 1, 0)
+        # tcd_raster.save(mask_path_tcd)
 
-def zonal_stats(input_folder):
-    aoi_list = list_files_in_directory(input_folder, '.shp')
+        # Conditional logic for Gain
+        gain_raster: object = arcpy.sa.Con(arcpy.Raster(gain_raster) > 0, 1, 0)
 
+        # Conditional logic for WHRC
+        whrc_raster = arcpy.sa.Con(arcpy.Raster(whrc_raster) > 0, 1, 0)
+
+        # Conditional logic for Mangrove
+        mangrove_raster = arcpy.sa.Con(arcpy.Raster(mangrove_raster) > 0, 1, 0)
+
+        # Add raster calculator logic...
+        output_times = arcpy.sa.Times(tcd_raster, whrc_raster)
+        output_plus = arcpy.sa.Plus(output_times, gain_raster)
+        output_plus_2 = arcpy.sa.Plus(output_plus, mangrove_raster)
+        output_raster = arcpy.sa.Con(arcpy.Raster(output_plus_2) > 0, 1, 0)
+
+        output_raster.save(mask_path_tcd_gain)
+
+    TDC_threshold = 30
+    gain = False
+
+    # Get a list of rasters in the raster folder
+    tcd_list = [os.path.join(tcd_tiles, f) for f in os.listdir(tcd_tiles) if f.endswith('.tif')]
+    print(tcd_list)
+
+    # Loop through each raster and calculate zonal statistics as table
+    for raster in tcd_list:
+        # Extract the name of the raster from the file path
+        raster_name = os.path.splitext(os.path.basename(raster))[0]
+        # print(raster_name)
+        tile_id = raster_name.split("_", 3)[3]
+        # print(tile_id)
+        gain_raster = os.path.join(gain_tiles, str(tile_id + '_gain.tif'))
+        mangrove_raster = os.path.join(mangrove_tiles, str(tile_id + '_mangrove_agb_t_ha_2000_rewindow.tif'))
+        whrc_raster = os.path.join(whrc_tiles, str(tile_id + '_t_aboveground_biomass_ha_2000.tif'))
+
+        mask_path_tcd = os.path.join(mask_tiles, str(tile_id + '_tcd.tif'))
+        mask_path_tcd_gain = os.path.join(mask_tiles, str(tile_id + '_tcd_gain.tif'))
+        print('Creating masks for ' + tile_id + ": \n")
+
+        # Conditional logic for TCD
+        tcd_raster = arcpy.sa.Con(arcpy.Raster(raster) > 30, 1, 0)
+        # tcd_raster.save(mask_path_tcd)
+
+        # Conditional logic for Gain
+        gain_raster: object = arcpy.sa.Con(arcpy.Raster(gain_raster) > 0, 0, 0)
+
+        # Conditional logic for WHRC
+        whrc_raster = arcpy.sa.Con(arcpy.Raster(whrc_raster) > 0, 1, 0)
+
+        # Conditional logic for Mangrove
+        mangrove_raster = arcpy.sa.Con(arcpy.Raster(mangrove_raster) > 0, 1, 0)
+
+        # Add raster calculator logic...
+        output_times = arcpy.sa.Times(tcd_raster, whrc_raster)
+        output_plus = arcpy.sa.Plus(output_times, gain_raster)
+        output_plus_2 = arcpy.sa.Plus(output_plus, mangrove_raster)
+        output_raster = arcpy.sa.Con(arcpy.Raster(output_plus_2) > 0, 1, 0)
+
+        output_raster.save(mask_path_tcd)
+
+def ZonalStats(input_folder: object) -> object:
+    # Get a list of shapefiles in the input folder
+    aoi_list = [os.path.join(input_folder, f) for f in os.listdir(input_folder) if f.endswith('.shp')]
+
+    print(aoi_list)
+    # Loop through each area of interest
     for aoi in aoi_list:
+        # Extract the name of the AOI from the file path
         aoi_name = os.path.splitext(os.path.basename(aoi))[0]
-        print(f"Now processing {aoi_name}: \n")
+        print("Now processing " + aoi_name + ": \n")
 
         if "IDN" in aoi_name:
-            raster_folder = os.path.join(arcpy.env.workspace, "Input", "00N_110E")
+            raster_folder = os.path.join(arcpy.env.workspace,"Input","00N_110E")
             output_folder = os.path.join(arcpy.env.workspace, "Outputs", "00N_110E")
         else:
             raster_folder = os.path.join(arcpy.env.workspace, "Input", "20N_20W")
             output_folder = os.path.join(arcpy.env.workspace, "Outputs", "20N_20W")
 
-        process_zonal_statistics(aoi, raster_folder, output_folder)
+        # Get a list of rasters in the raster folder
+        raster_list = [os.path.join(raster_folder, f) for f in os.listdir(raster_folder) if f.endswith('.tif')]
 
-def zonal_stats_masked(input_folder):
-    aoi_list = list_files_in_directory(input_folder, '.shp')
-
-    for aoi in aoi_list:
-        aoi_name = os.path.splitext(os.path.basename(aoi))[0]
-        print(f"Now processing {aoi_name}: \n")
-
-        if "IDN" in aoi_name:
-            raster_folder = os.path.join(arcpy.env.workspace, "Input", "00N_110E")
-            output_folder = os.path.join(arcpy.env.workspace, "Outputs", "00N_110E")
-            mask_tiles = os.path.join(arcpy.env.workspace, "Mask", "Mask", "00N_110E")
-        else:
-            raster_folder = os.path.join(arcpy.env.workspace, "Input", "20N_20W")
-            output_folder = os.path.join(arcpy.env.workspace, "Outputs", "20N_20W")
-            mask_tiles = os.path.join(arcpy.env.workspace, "Mask", "Mask", "20N_20W")
-
-        raster_list = list_files_in_directory(raster_folder, '.tif')
-        mask_list = list_files_in_directory(mask_tiles, '.tif')
-
+        # Loop through each raster and calculate zonal statistics as table
         for raster in raster_list:
-            raster_obj = arcpy.Raster(raster)
+            # Extract the name of the raster from the file path
+            raster_name = os.path.splitext(os.path.basename(raster))[0]
+            print('Calculating zonal statistics for ' + raster_name + ": \n")
 
-            for mask in mask_list:
-                mask_obj = arcpy.Raster(mask)
+            # Create a name for the output table by concatenating the AOI name and raster name
+            output_name = "{}_{}.dbf".format(aoi_name, raster_name)
+            output_path = os.path.join(output_folder, output_name)
 
-                # Check if spatial references and extents are the same
-                if (raster_obj.extent == mask_obj.extent and
-                        raster_obj.spatialReference.name == mask_obj.spatialReference.name):
+            # Calculate zonal statistics as table for the current raster and AOI
+            arcpy.gp.ZonalStatisticsAsTable_sa(aoi, "GID_0", raster, output_path, "DATA", "SUM")
 
-                    masked_raster = arcpy.sa.Times(raster_obj, mask_obj)
-                    # ... [rest of your code] ...
-                else:
-                    print(f"Spatial references or extents do not match for {raster} and {mask}")
+            # convert each output to csv
+            csv_file = "{}_{}.csv".format(aoi_name, raster_name)
+            arcpy.TableToTable_conversion(output_path, output_folder, csv_file)
 
-def process_annual_zonal_stats(aoi, raster_folder, output_folder):
-    raster_list = list_files_in_directory(raster_folder, 'emis.tif')
+def ZonalStatsMasked(input_folder: str) -> object:
+    """
 
-    for raster in raster_list:
-        raster_name = os.path.splitext(os.path.basename(raster))[0]
-        print(f'Calculating zonal statistics for {raster_name}: \n')
+    :rtype: object
+    """
+    input_folder = os.path.join(arcpy.env.workspace, "AOIS")
+    #TODO: ADD INPUT FOLDER AS AN ARGUMENT RATHER THAN HARD CODING IT HERE
 
-        output_name = f"{os.path.splitext(os.path.basename(aoi))[0]}_{raster_name}.dbf"
-        output_path = os.path.join(output_folder, output_name)
+    # Get a list of shapefiles in the input folder
+    aoi_list = [os.path.join(input_folder, f) for f in os.listdir(input_folder) if f.endswith('.shp')]
+    print(aoi_list)
 
-        arcpy.gp.ZonalStatisticsAsTable_sa(aoi, "Value", raster, output_path, "DATA", "SUM")
-        csv_file = f"{os.path.splitext(os.path.basename(aoi))[0]}_{raster_name}.csv"
-        arcpy.TableToTable_conversion(output_path, output_folder, csv_file)
-
-def zonal_stats_annualized(input_folder):
-    aoi_list = list_files_in_directory(input_folder, 'clip.tif')
-
+    # Loop through each area of interest
     for aoi in aoi_list:
+        # Extract the name of the AOI from the file path
         aoi_name = os.path.splitext(os.path.basename(aoi))[0]
-        print(f"Now processing {aoi_name}: \n")
+        print("Now processing " + aoi_name + ": \n")
 
         if "IDN" in aoi_name:
-            raster_folder = os.path.join(arcpy.env.workspace, "Input", "00N_110E")
+            raster_folder = os.path.join(arcpy.env.workspace,"Input","00N_110E")
+            output_folder = os.path.join(arcpy.env.workspace, "Outputs", "00N_110E")
+            mask_tiles = os.path.join(arcpy.env.workspace, "Mask","Mask","00N_110E")
+        else:
+            raster_folder = os.path.join(arcpy.env.workspace, "Input", "20N_20W")
+            output_folder = os.path.join(arcpy.env.workspace, "Outputs", "20N_20W")
+            mask_tiles = os.path.join(arcpy.env.workspace, "Mask", "Mask","20N_20W")
+
+        # Get a list of rasters in the raster folder
+        raster_list = [os.path.join(raster_folder, f) for f in os.listdir(raster_folder) if f.endswith('.tif')]
+
+        # Loop through each raster and calculate zonal statistics as table
+        for raster in raster_list:
+            # Extract the name of the raster from the file path
+            raster_name = os.path.splitext(os.path.basename(raster))[0]
+            #tile_id = str(raster_name.split("_", 1)[0]) + "_" + str(raster_name.split("_", 2)[1])
+
+            print('Calculating zonal statistics for ' + raster_name + ": \n")
+
+            mask_list = [os.path.join(mask_tiles, f) for f in os.listdir(mask_tiles) if f.endswith('.tif')]
+            for mask in mask_list:
+                mask_path = os.path.splitext(os.path.basename(mask))[0]
+                mask_name = mask_path.split("_", 2)[2]
+
+                # # Create a name for the output table by concatenating the AOI name and raster name
+                output_name = "{}_{}.dbf".format(aoi_name, str(raster_name) + "_" + str(mask_name))
+                output_path = os.path.join(output_folder, output_name)
+
+                masked_raster = arcpy.sa.Times(raster, mask)
+                arcpy.gp.ZonalStatisticsAsTable_sa(aoi, "GID_0", masked_raster, output_path, "DATA", "SUM")
+
+                # convert each output to csv
+                csv_file = "{}_{}.csv".format(aoi_name, str(raster_name) + "_" + str(mask_name))
+                arcpy.TableToTable_conversion(output_path, output_folder, csv_file)
+
+def ZonalStatsAnnualized(input_folder):
+    # Get a list of shapefiles in the input folder
+    input_folder = os.path.join(arcpy.env.workspace, "TCL")
+    # TODO: ADD INPUT FOLDER AS AN ARGUMENT RATHER THAN HARD CODING IT HERE
+    aoi_list = [os.path.join(input_folder, f) for f in os.listdir(input_folder) if f.endswith('clip.tif')]
+
+    print(aoi_list)
+
+    # Loop through each area of interest
+    for aoi in aoi_list:
+        # Extract the name of the AOI from the file path
+        aoi_name = os.path.splitext(os.path.basename(aoi))[0]
+        print("Now processing " + aoi_name + ": \n")
+
+        if "IDN" in aoi_name:
+            raster_folder = os.path.join(arcpy.env.workspace,"Input","00N_110E")
             output_folder = os.path.join(arcpy.env.workspace, "Outputs", "Annual")
         else:
             raster_folder = os.path.join(arcpy.env.workspace, "Input", "20N_20W")
-            output_folder = os.path.join(arcpy.env.workspace, "Outputs", "Annual")
+            output_folder = os.path.join(arcpy.env.workspace, "Outputs","Annual")
 
-        process_annual_zonal_stats(aoi, raster_folder, output_folder)
+        # Get a list of rasters in the raster folder
+        raster_list = [os.path.join(raster_folder, f) for f in os.listdir(raster_folder) if "emis" in f and f.endswith('tif')]
 
-def load_and_process_csv(file_path, file_name):
-    csv_df = pd.read_csv(file_path)
+        # Loop through each raster and calculate zonal statistics as table
+        for raster in raster_list:
+            # Extract the name of the raster from the file path
+            raster_name = os.path.splitext(os.path.basename(raster))[0]
+            print('Calculating zonal statistics for ' + raster_name + ": \n")
 
-    # Assigning attributes based on file name
-    csv_df["Name"] = file_name
-    csv_df["Type"] = "gross emissions" if "emis" in file_name else "gross removals" if "removals" in file_name else "net flux"
-    csv_df["Extent"] = "forest extent" if "forest" in file_name else "full extent"
-    csv_df["Mask"] = "tcd and gain" if "tcd_gain" in file_name else "tcd" if "_tcd." in file_name else "no mask"
+            # Create a name for the output table by concatenating the AOI name and raster name
+            output_name = "{}_{}.dbf".format(aoi_name, raster_name)
+            output_path = os.path.join(output_folder, output_name)
 
-    # Dropping unnecessary columns
-    csv_df.drop(['OID_', 'COUNT', 'AREA'], axis=1, inplace=True)
-    return csv_df
+            # Calculate zonal statistics as table for the current raster and AOI
+            arcpy.gp.ZonalStatisticsAsTable_sa(aoi, "Value", raster, output_path, "DATA", "SUM")
 
-def zonal_stats_clean(input_folders):
+            # convert each output to csv
+            csv_file = "{}_{}.csv".format(aoi_name, raster_name)
+            arcpy.TableToTable_conversion(output_path, output_folder, csv_file)
+
+def ZonalStatsClean(input_folder) -> object:
+    # Initialize an empty data frame to store the data
+
+
+    folder1 = os.path.join(arcpy.env.workspace,"Outputs","00N_110E")
+    folder2 = os.path.join(arcpy.env.workspace,"Outputs","20N_20W")
+    folder3 = os.path.join(arcpy.env.workspace,"Outputs","Annual")
+
     df = pd.DataFrame()
 
-    for folder in input_folders:
-        for file in os.listdir(folder):
-            if file.endswith(".csv"):
-                file_path = os.path.join(folder, file)
-                csv_df = load_and_process_csv(file_path, file)
-                df = pd.concat([df, csv_df], axis=0)
+    # Loop through the files in folder1
+    for file in os.listdir(folder1):
+        if file.endswith(".csv"):
+            # Load the csv file into a pandas data frame
+            csv_df = pd.read_csv(os.path.join(folder1, file))
 
+            # Rename the "sum" field to the name of the file
+            csv_df["Name"] = file
+
+
+            #Define type of calc
+            if "emis" in file:
+                type = "gross emissions"
+            elif "removals" in file:
+                type = "gross removals"
+            else:
+                type = "net flux"
+            csv_df["Type"] = type
+
+            #Define extent of calc
+            if "forest_extent" in file:
+                extent = "forest extent"
+            else:
+                extent = "full extent"
+            csv_df["Extent"] = extent
+
+            # Define mask of calc
+            if "tcd_gain" in file:
+                mask = "tcd and gain"
+            elif "_tcd." in file:
+                mask = "tcd"
+            else:
+                mask = "no mask"
+            csv_df["Mask"] = mask
+
+            # Drop all other fields
+            assert isinstance(csv_df, object)
+            csv_df.drop([ 'OID_', 'COUNT', 'AREA'], axis=1, inplace=True)
+
+            # Append the data to the main data frame
+            df = pd.concat([df, csv_df], axis=0)
+
+    # Loop through the files in folder2
+    for file in os.listdir(folder2):
+        if file.endswith(".csv"):
+            # Load the csv file into a pandas data frame
+            csv_df = pd.read_csv(os.path.join(folder2, file))
+
+            # Rename the "sum" field to the name of the file
+            csv_df["Name"] = file
+            #Define type of calc
+            if "emis" in file:
+                type = "gross emissions"
+            elif "removals" in file:
+                type = "gross removals"
+            else:
+                type = "net flux"
+            csv_df["Type"] = type
+
+            #Define extent of calc
+            if "forest" in file:
+                extent = "forest extent"
+            else:
+                extent = "full extent"
+            csv_df["Extent"] = extent
+
+            # Define mask of calc
+            if "tcd_gain" in file:
+                mask = "tcd and gain"
+            elif "_tcd." in file:
+                mask = "tcd"
+            else:
+                mask = "no mask"
+            csv_df["Mask"] = mask
+
+            # Drop all other fields
+            assert isinstance(csv_df, object)
+            csv_df.drop([ 'OID_', 'COUNT', 'AREA'], axis=1, inplace=True)
+
+            # Append the data to the main data frame
+            df = pd.concat([df, csv_df], axis=0)
+
+    # Print the resulting data frame to check the data
     print(df)
 
-    output_path = os.path.join(arcpy.env.workspace, "Outputs", "CSV", "output.csv")
+    # define the output location
+    output_path = os.path.join(arcpy.env.workspace, "Outputs", "CSV","output.csv")
+
+    # Export the data frame as a CSV file
     df.to_csv(output_path, index=False)
 
-#todo solve mask output problem
-#todo solve annualized output and final csv problem
+#TODO: Export dataframe for annualized results
 
 # Downloads individual tiles from s3
     # source = source file on s3
     # dest = where to download
     # pattern = pattern for file name
-def s3_file_download(source, dest, pattern):
-
+def s3_file_download(source, dest, pattern=''):
     # Retrieves the s3 directory and name of the tile from the full path name
     dir = get_tile_dir(source)
     file_name = get_tile_name(source)
@@ -203,7 +423,7 @@ def s3_file_download(source, dest, pattern):
 
     # Special download procedures for tree cover gain because the tiles have no pattern, just an ID.
     # Tree cover gain tiles are renamed as their downloaded to get a pattern added to them.
-    if dir == cn.gain_dir[:-1]: # Delete last character of gain_dir because it has the terminal / while dir does not have terminal /
+    if dir == cn.gain_s3_path[:-1]: # Delete last character of gain_dir because it has the terminal / while dir does not have terminal /
         local_file_name = f'{tile_id}_{pattern}.tif'
         print(f'Option 1: Checking if {local_file_name} is already downloaded...')
         if os.path.exists(os.path.join(dest, local_file_name)):
@@ -214,12 +434,15 @@ def s3_file_download(source, dest, pattern):
             print(f'Option 2: Checking for tile {source} on s3...')
 
             # If the tile isn't already downloaded, download is attempted
-            source = os.path.join(dir, file_name)
+            # source = os.path.join(dir, file_name)
+            source = f'{dir}/{file_name}'
+            local_folder = os.path.join(dest, local_file_name)
 
             # cmd = ['aws', 's3', 'cp', source, dest, '--no-sign-request', '--only-show-errors']
-            cmd = ['aws', 's3', 'cp', source, f'{dest}{local_file_name}',
+            cmd = ['aws', 's3', 'cp', source, local_folder,
                    '--request-payer', 'requester', '--only-show-errors']
             log_subprocess_output_full(cmd)
+
             if os.path.exists(os.path.join(dest, local_file_name)):
                 print_log(f'  Option 2 success: Tile {source} found on s3 and downloaded', "\n")
                 return
@@ -238,7 +461,8 @@ def s3_file_download(source, dest, pattern):
             print_log(f'Option 2: Checking for tile {source} on s3...')
 
             # If the tile isn't already downloaded, download is attempted
-            source = os.path.join(dir, file_name)
+            #source = os.path.join(dir, file_name)
+            source = f'{dir}/{file_name}'
 
             # cmd = ['aws', 's3', 'cp', source, dest, '--no-sign-request', '--only-show-errors']
             cmd = ['aws', 's3', 'cp', source, dest, '--only-show-errors']
@@ -249,18 +473,23 @@ def s3_file_download(source, dest, pattern):
             else:
                 print_log(f'Option 2 failure: Tile {source} not found on s3. Tile not found but it seems it should be. Check file paths and names.', "\n")
 
-def s3_flexible_download(s3_dir, s3_pattern, local_dir, local_pattern, tile_id_list):
+def s3_flexible_download(tile_id_list, s3_dir, s3_pattern, local_dir, local_pattern = ''):
 
     # Creates a full download name (path and file)
     for tile_id in tile_id_list:
-        if s3_pattern in [cn.pattern_tcd, cn.pattern_pixel_area, cn.pattern_loss]:
+        if s3_pattern in [cn.tcd_s3_pattern, cn.loss_s3_pattern]:
             source = f'{s3_dir}{s3_pattern}_{tile_id}.tif'
-        elif s3_pattern in [cn.pattern_data_lake]:
+        elif s3_pattern in [cn.gain_s3_pattern]:
             source = f'{s3_dir}{tile_id}.tif'
         else:  # For every other type of tile
             source = f'{s3_dir}{tile_id}_{s3_pattern}.tif'
 
-        s3_file_download(source, local_dir, local_pattern)
+        if s3_pattern in [cn.gross_emis_forest_extent_s3_pattern, cn.gross_emis_full_extent_s3_pattern, cn.gross_removals_forest_extent_s3_pattern, cn.gross_removals_full_extent_s3_pattern, cn.netflux_forest_extent_s3_pattern, cn.netflux_full_extent_s3_pattern]:
+            dir = os.path.join(local_dir, tile_id)
+        else:
+            dir = local_dir
+
+        s3_file_download(source, dir, local_pattern)
 
 # Gets the directory of the tile
 def get_tile_dir(tile):
@@ -308,3 +537,13 @@ def print_log(*args):
 
     # Prints to console
     print("LOG: " + full_statement)
+
+def folder_check(folder):
+        if os.path.isdir(folder):
+            print(f"    Option 1 success: {folder} exists.")
+        else:
+            os.makedirs(folder)
+            if os.path.isdir(folder):
+                print(f"    Option 2 success: {folder} successfully created.")
+            else:
+                raise Exception(f"  Option 2 failure: {folder} could not be created.")
